Added new ESP8266-as-Arduino project (last meeting + spring break)
ESP8266_AS_ARDUINO_PROTOTYPE is the latest project that we will likely
be using. We determined that it was best to use the ESP8266 directly,
rather than interfacing it with the Arduino (which was cumbersome).

From the previous meeting, we got WiFi, NTP time updating via
WiFi, and timers to work.

During spring break, this was expanded to fixing the timers (uint32_t
instead of uint8_t for storage and timer returns), microphone sensing
(via analogRead + tuning), and WiFi enabling/disabling (with
modifications to NTP updater for enabling WiFi if it's off).