Use local variable for timer instead + update API to encourage this
malloc() is a bad idea for embedded platforms (though I didn't know
this at the time I originally wrote this). As such, it has been
removed from the main INO in favor of a local function variable
(stored in the stack). That way, it's faster and there's much less
chance for memory issues (malloc(), leaks, etc.).

Documentation is updated to reflect this sentiment. A small fix has
also been added to ensure that the pointer is NULL when malloc()
fails.